requirements.txt
django
langchain
python-dotenv
openai
numpy
faiss-cpu
pypdf2
django-restframework
djangorestframework-simplejwt
django-cors-headers
pyrebase4
gensim
